libs/unit: Fix LVGL win switch width float issue.

+++ m5stack/libs/m5ui/win.py
@@ -7,7 +7,7 @@ import lvgl as lv
 import m5ui
 
 try:
-    DPI = lv.DPI_DEF / 3
+    DPI = int(lv.DPI_DEF / 3)
 except Exception:
     DPI = 80
